requirements.txt
absl-py==2.1.0
aioice==0.9.0
aiortc==1.8.0
altair==5.2.0
attrs==23.2.0
av==11.0.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
dnspython==2.6.1
flatbuffers==20181003210633
fonttools==4.50.0
gitdb==4.0.11
GitPython==3.1.42
google-crc32c==1.5.0
idna==3.6
ifaddr==0.2.0
importlib-metadata==7.0.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
mediapipe==0.10.9
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.1
pillow==10.2.0
protobuf==3.20.3
pyarrow==15.0.0
pycparser==2.21
pycups==2.0.1
pydeck==0.8.1b0
pyee==11.1.0
pygame==2.5.2
Pygments==2.17.2
pylibsrtp==0.10.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
pytz==2024.1
qrcode==7.4.2
referencing==0.33.0
reportlab==4.1.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
sounddevice==0.4.6
streamlit==1.32.2
streamlit-webrtc==0.47.6
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
validators==0.22.0
watchdog==4.0.0
zipp==3.17.0
